Add all the changes, too lazy to separate
Added support for height maps in renderer, added simplex noise based height map generation,
water rendering, a mildly accurate (but currently broken) moon, proper sunset/sunrise to skybox
using texture lookups, procedural stars, made sure to clean up interleaved buffer before creating
buffers for depth map in *batch classes, renamed/moved some classes

+++ app/src/main/java/com/sonicmax/bloodrogue/engine/collisions/FieldOfVisionCalculator.java
@@ -12,7 +12,7 @@ public class FieldOfVisionCalculator {
     private final String LOG_TAG = this.getClass().getSimpleName();
     private double darknessFactor; // Lower value = darker lighting. Has to be > 0
 
-    private long[][] mapGrid;
+    private int[][] mapGrid;
     private ArrayList<Long>[][] objectGrid;
     private ComponentManager componentManager;
     private int startX;
@@ -33,7 +33,7 @@ public class FieldOfVisionCalculator {
         return this.visitedTiles;
     }
 
-    public void setValues(long[][] mapGrid, ArrayList<Long>[][] objectGrid, int x, int y, int radius) {
+    public void setValues(int[][] mapGrid, ArrayList<Long>[][] objectGrid, int x, int y, int radius) {
 
         this.mapGrid = mapGrid;
         this.objectGrid = objectGrid;
